Testing | Small test

+++ src/main/java/com/example/springLoan_18112018/repository/SettingRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SettingRepository extends JpaRepository<Setting, Long> {
+public interface SettingRepository extends JpaRepository<Setting, Integer> {
 }
